refacto(compare)r: compare module restrucurization for cleaner strucutre
